Fixes
BFV/CKKS : ciphertext marshaling now properly works when the target is a ciphertext that isnt a bfv or ckks element.

BFV/CKKS : API change on key-generator, Gen is now a prefix to indicate that there is sampling, while New is a prefixe to indicate the creation of an empty struct.