making the evaluation keys independant of the OperandQP type
As for the public keys, I think it is better to keep evaluation key types
as simple as possible. The "Operand/OperandQP" types are more adapted to
the "data" path, i.e. as operands in a circuit. One obvious exemple is that
there is no point for keys to have the metadata of a ciphertext.Also, we'll
have easier time designing the evaluation logic and evolving the Operand
types if the keys do not depend on them.